Crate sp_evm_precompiles

EVM precompile primitives

Enums

PrecompileName

Type Aliases

Precompiles

The PrecompileSet installed in the EVM Domain runtime. The following distribution has been decided for the precompiles 0-1023: Ethereum Mainnet Precompiles 1024-2047 Precompiles that are not in Ethereum Mainnet, and are not Subspace specific 2048-4095 Subspace specific precompiles