sp_evm_precompiles

Type Alias Precompiles

pub type Precompiles<R> = PrecompileSetBuilder<R, (PrecompilesInRangeInclusive<(AddressU64<1>, AddressU64<4095>), (PrecompileAt<AddressU64<1>, ECRecover, (AcceptDelegateCall, CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<2>, Sha256, (AcceptDelegateCall, CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<3>, Ripemd160, (AcceptDelegateCall, CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<4>, Identity, (AcceptDelegateCall, CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<5>, Modexp, (AcceptDelegateCall, CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<1024>, Sha3FIPS256, (CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<1025>, ECRecoverPublicKey, (CallableByContract, CallableByPrecompile)>, PrecompileAt<AddressU64<2048>, TransporterPrecompile<R>, (CallableByContract, CallableByPrecompile)>)>,)>;

The PrecompileSet installed in the EVM Domain runtime. The following distribution has been decided for the precompiles 0-1023: Ethereum Mainnet Precompiles 1024-2047 Precompiles that are not in Ethereum Mainnet, and are not Subspace specific 2048-4095 Subspace specific precompiles

Aliased Type

pub struct Precompiles<R> { /* private fields */ }